feat(configuration-core): introduce configuration-core package for feature flag management
