+++ packages/argent-cli/src/index.ts
@@ -1,13 +1,14 @@
 export { run, type RunCommandOptions } from "./run.js";
 export { tools, type ToolsCommandOptions } from "./tools.js";
 export { server } from "./server.js";
+export { enable, disable, flags } from "./flags.js";
+// Backward-compat re-export: the flag primitives now live in
+// @argent/configuration-core, but @argent/cli's public surface keeps exposing
+// them so existing importers (and the publish bundle) are unaffected.
 export {
-  enable,
-  disable,
-  flags,
   isFlagEnabled,
   getFlagDefinition,
   FLAG_REGISTRY,
   type FlagScope,
   type FlagDefinition,
-} from "./flags.js";
+} from "@argent/configuration-core";
