+++ packages/configuration-core/src/index.ts
@@ -1,0 +1,13 @@
+export {
+  FLAG_REGISTRY,
+  getFlagDefinition,
+  resolveProjectRoot,
+  getFlagsPath,
+  readFlags,
+  setFlag,
+  unsetFlag,
+  isFlagEnabled,
+  type FlagScope,
+  type FlagDefinition,
+  type FlagsPathOptions,
+} from "./flags.js";
